PP-2883 Refactor the package structure
- Refactor the package structure to be "package by feature, not layer"

More information: http://www.javapractices.com/topic/TopicAction.do?Id=205
