+++ pom.xml
@@ -18,7 +18,7 @@
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
         <dropwizard.version>1.2.0</dropwizard.version>
-        <mainClass>uk.gov.pay.directdebit.app.DirectDebitConnectorApp</mainClass>
+        <mainClass>uk.gov.pay.directdebit.DirectDebitConnectorApp</mainClass>
 
         <liquibase.version>3.5.3</liquibase.version>
         <postgresql.version>42.1.4</postgresql.version>
