+++ src/main/java/uk/gov/pay/directdebit/app/filters/LoggingFilter.java
@@ -42,7 +42,7 @@ public class LoggingFilter implements Filter {
         try {
             filterChain.doFilter(servletRequest, servletResponse);
         } catch (Throwable throwable) {
-            logger.error("Exception - adminusers request - " + requestURL + " - exception - " + throwable.getMessage(), throwable);
+            logger.error("Exception - direct debit connector request - " + requestURL + " - exception - " + throwable.getMessage(), throwable);
         } finally {
             logger.info(format("%s to %s ended - total time %dms", requestMethod, requestURL,
                     stopwatch.elapsed(TimeUnit.MILLISECONDS)));
