+++ src/main/java/uk/gov/pay/directdebit/webhook/gocardless/config/GoCardlessFactory.java
@@ -1,9 +1,9 @@
-package uk.gov.pay.directdebit.app.config;
+package uk.gov.pay.directdebit.webhook.gocardless.config;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.gocardless.GoCardlessClient;
 import org.apache.commons.lang3.StringUtils;
-import uk.gov.pay.directdebit.app.core.WebhookVerifier;
+import uk.gov.pay.directdebit.webhook.gocardless.support.WebhookVerifier;
 
 import javax.validation.constraints.NotNull;
