+++ src/main/java/uk/gov/pay/directdebit/webhook/gocardless/exception/InvalidWebhookExceptionMapper.java
@@ -1,4 +1,4 @@
-package uk.gov.pay.directdebit.app.exception;
+package uk.gov.pay.directdebit.webhook.gocardless.exception;
 
 import org.slf4j.Logger;
 import uk.gov.pay.directdebit.app.logger.PayLoggerFactory;
@@ -10,13 +10,19 @@ public class InvalidWebhookExceptionMapper implements ExceptionMapper<InvalidWeb
 
     private static final Logger LOGGER = PayLoggerFactory.getLogger(InvalidWebhookExceptionMapper.class);
 
+    /**
+     * Map an exception to a {@link javax.ws.rs.core.Response}.
+     * Webhooks with an invalid signature must return a "498 Token Invalid" error.
+     *
+     * @param exception the exception to map to a response.
+     * @return a response mapped from the supplied exception.
+     * @see <a href="https://developer.gocardless.com/api-reference/#webhooks-overview">https://developer.gocardless.com/api-reference/#webhooks-overview</a>
+     */
     @Override
     public Response toResponse(InvalidWebhookException exception) {
         // Log an error and possibly do SumoLogic alert
         LOGGER.error(exception.getMessage());
 
-        // Webhooks with an invalid signature must return a "498 Token Invalid" error.
-        // see https://developer.gocardless.com/api-reference/#webhooks-overview
         return Response.status(498).build();
     }
