+++ src/test/java/uk/gov/pay/directdebit/webhook/gocardless/support/WebhookVerifierTest.java
@@ -1,8 +1,8 @@
-package uk.gov.pay.directdebit.app.core;
+package uk.gov.pay.directdebit.webhook.gocardless.support;
 
 import com.google.common.io.Resources;
 import org.junit.Test;
-import uk.gov.pay.directdebit.app.exception.InvalidWebhookException;
+import uk.gov.pay.directdebit.webhook.gocardless.exception.InvalidWebhookException;
 
 import static com.google.common.base.Charsets.UTF_8;
 
@@ -22,7 +22,7 @@ public class WebhookVerifierTest {
     @Test
     public void shouldVerifyWebhookWithCorrectSignature() throws Exception {
         String signature = "4d48a688e8bd6c313e3eecc78fa55b3e4ae23c65e70cf35038010f47742fb670";
-        String body = Resources.toString(Resources.getResource("webhook.json"), UTF_8);
+        String body = Resources.toString(Resources.getResource("gocardless-webhook.json"), UTF_8);
 
         verifier.verify(body, signature);
     }
@@ -30,7 +30,7 @@ public class WebhookVerifierTest {
     @Test(expected = InvalidWebhookException.class)
     public void shouldThrowForWebhookWithIncorrectSignature() throws Exception {
         String signature = "4d48a688e8bd6c313e3eecc78fa55b3e4ae23c65e70cf35038010f47742fb671";
-        String body = Resources.toString(Resources.getResource("webhook.json"), UTF_8);
+        String body = Resources.toString(Resources.getResource("gocardless-webhook.json"), UTF_8);
 
         verifier.verify(body, signature);
     }
